requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==2.1.15
django-environ==0.8.1
django-filter==21.1
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
djangorestframework==3.12.4
importlib-metadata==4.8.2
Markdown==3.3.6
pycodestyle==2.8.0
pyodbc==4.0.32
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
zipp==3.6.0
